requirements.txt
# Core face recognition libraries
# Using dlib wheel for easier installation
dlib==19.24.2
face-recognition==1.3.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# Web API
flask==3.1.0
werkzeug==3.1.3
requests==2.32.5
gunicorn==23.0.0

# Optional but recommended
click==8.1.7  # For CLI interface

# Required for Python 3.12+
setuptools>=65.5.1
